fixed timout issue on hoymiels inverter and removed currently workaround, also added Connection statistics for network invertes (implemented only HOymiles W-Series)

+++ lib/DeyeSun/src/inverters/DeyeInverter.h
@@ -73,7 +73,7 @@ private:
     
     String filterReceivedResponse(size_t length);
 
-    uint16_t getInternalPollTime();
+    uint64_t getInternalPollTime();
 
     bool resolveHostname();
     void swapBuffers(bool fullData);
